removed redundant state and actions from useAudioPlayerData.
