+++ .env.development
@@ -1,1 +1,1 @@
-VITE_CSP=default-src 'self'; connect-src 'self' http://localhost:8000; script-src 'self'; style-src 'self' 'unsafe-inline'; font-src 'self'; img-src 'self' *;
+VITE_CSP=default-src 'self'; connect-src 'self' http://localhost:8000; script-src 'self'; style-src 'self' 'unsafe-inline'; font-src 'self'; img-src 'self' *; media-src 'self' http://localhost:8000;
