+++ src/shared/model/stores/playlistStore.ts
@@ -7,7 +7,6 @@ const playlistStore = create<PlaylistStore>()((set) => ({
   isPlaying: false,
   direction: "forward",
   isInitialized: false,
-  audioControl: null,
   queue: [],
   actions: {
     setTracks: (tracks) => {
@@ -17,7 +16,6 @@ const playlistStore = create<PlaylistStore>()((set) => ({
     setIsPlaying: (isPlaying) => set({ isPlaying }),
     setDirection: (direction) => set({ direction }),
     setIsInitialized: (isInitialized) => set({ isInitialized }),
-    setAudioControl: (audioControl) => set({ audioControl }),
     pushTrackToQueue: (track) => {
       set((state) => ({
         queue: [...state.queue, track],
@@ -43,6 +41,5 @@ export const usePlaylistCurrentTrackIndex = () => playlistStore((state) => state
 export const usePlaylistIsPlaying = () => playlistStore((state) => state.isPlaying);
 export const usePlaylistDirection = () => playlistStore((state) => state.direction);
 export const usePlaylistIsInitialized = () => playlistStore((state) => state.isInitialized);
-export const usePlaylistAudioControl = () => playlistStore((state) => state.audioControl);
 export const usePlaylistQueue = () => playlistStore((state) => state.queue);
 export const usePlaylistActions = () => playlistStore((state) => state.actions);
