+++ src/shared/model/types/playerStore.ts
@@ -3,6 +3,7 @@ interface PlayerState {
   duration: number;
   currentProgress: number;
   buffered: number;
+  audioControl: HTMLAudioElement | null;
 }
 
 interface PlayerActions {
@@ -10,8 +11,9 @@ interface PlayerActions {
   setDuration: (duration: number) => void;
   setCurrentProgress: (progress: number) => void;
   setBuffered: (buffered: number) => void;
+  setAudioControl: (audioControl: HTMLAudioElement | null) => void;
 }
 
 export interface PlayerStore extends PlayerState {
-    actions: PlayerActions;
+  actions: PlayerActions;
 }
